+++ CHANGELOG.md
@@ -3,6 +3,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### ✅ Added
+- Add the `makeAttachmentTextView` method to ViewFactory [#1013](https://github.com/GetStream/stream-chat-swiftui/pull/1013)
+
 ### 🔄 Changed
 
 # [4.90.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.90.0)